refactor(js): route components in isolated files
