+++ src/App.js
@@ -1,7 +1,10 @@
 import Navbar from './components/Navbar';
 import Footer from './components/Footer';
-import { BrowserRouter, Routes, Route, Navigate, useParams } from 'react-router-dom';
-import { useEffect, useState } from 'react';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { Home } from './routes/Home';
+import { About } from './routes/About';
+import { MoviePage } from './routes/MoviePage';
+import { NotFound } from './routes/NotFound';
 
 
 const routes = [
@@ -27,66 +30,6 @@ const routes = [
     exact: false,
   },
 ]
-
-// TODO : preliminary implementation of page/route components
-function Home() {
-  return (
-    <div>
-      <h1>Home</h1>
-    </div>
-  );
-}
-
-function About() {
-  return (
-    <div>
-      <h1>About</h1>
-    </div>
-  );
-}
-
-function MoviePage() {
-  const { idFilm } = useParams();
-
-  return (
-    <div>
-      <h1>Movie Page {idFilm}</h1>
-    </div>
-  );
-}
-
-function NotFound() {
-  const timer = 5;
-  const [isRedirecting, setIsRedirecting] = useState(false);
-  const [countdown, setCountdown] = useState(timer);
-
-  useEffect(() => {
-    /* NOTE : To redirect after X seconds of countdown */
-    setTimeout(() => {
-      setIsRedirecting(true)
-    }, timer * 1_000)
-  }, [])
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setCountdown((prevCountdown) => prevCountdown - 1);
-    }, 1_000);
-
-    return () => clearInterval(interval);
-  })
-
-  if (isRedirecting) {
-    return <Navigate to="/" />
-  }
-
-
-  return (
-    <div>
-      <h1>404 Not Found</h1>
-      <p>Redirecting in {countdown} seconds...</p>
-    </div>
-  );
-}
 
 function App() {
   return (
