+++ src/routes/About.jsx
@@ -1,0 +1,7 @@
+export function About() {
+	return (
+		<div>
+			<h1>About</h1>
+		</div>
+	);
+}
