+++ src/routes/Home.jsx
@@ -1,0 +1,7 @@
+export function Home() {
+	return (
+		<div>
+			<h1>Home</h1>
+		</div>
+	);
+}
